move Composer and Ostree helpers to separate pkg
